*** empty log message ***

+++ org.eclipse.jdt.core/compiler/org/eclipse/jdt/internal/compiler/classfmt/ClassFileStruct.java
@@ -40,7 +40,14 @@ public int i4At(int relativeOffset) {
 }
 public long i8At(int relativeOffset) {
 	int position = relativeOffset + structOffset;
-	return (((long) (reference[position++] & 0xFF)) << 56) + (((long) (reference[position++] & 0xFF)) << 48) + (((long) (reference[position++] & 0xFF)) << 40) + (((long) (reference[position++] & 0xFF)) << 32) + (((long) (reference[position++] & 0xFF)) << 24) + (((long) (reference[position++] & 0xFF)) << 16) + (((long) (reference[position++] & 0xFF)) << 8) + ((long) (reference[position++] & 0xFF));
+	return (((long) (reference[position++] & 0xFF)) << 56) 
+					+ (((long) (reference[position++] & 0xFF)) << 48) 
+					+ (((long) (reference[position++] & 0xFF)) << 40) 
+					+ (((long) (reference[position++] & 0xFF)) << 32) 
+					+ (((long) (reference[position++] & 0xFF)) << 24) 
+					+ (((long) (reference[position++] & 0xFF)) << 16) 
+					+ (((long) (reference[position++] & 0xFF)) << 8) 
+					+ (reference[position++] & 0xFF);
 }
 public static String printTypeModifiers(int modifiers) {
